@@ -10,23 +10,8 @@
     <groupId>org.example</groupId>
     <artifactId>IM-relationship-server</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>IM-group-server</name>
-    <description>IM-group-server</description>
-    <url/>
-    <licenses>
-        <license/>
-    </licenses>
-    <developers>
-        <developer/>
-    </developers>
-    <scm>
-        <connection/>
-        <developerConnection/>
-        <tag/>
-        <url/>
-    </scm>
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
     </properties>
     <dependencies>
         <!-- Spring Boot core -->
@@ -45,6 +30,18 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-neo4j</artifactId>
+        </dependency>
+
+        <!-- PostgreSQL & MyBatis-Plus（用于组织架构 sys_* 表） -->
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <scope>runtime</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.baomidou</groupId>
+            <artifactId>mybatis-plus-boot-starter</artifactId>
         </dependency>
 
         <!-- Service discovery (Nacos) -->
@@ -72,10 +69,17 @@
             <optional>true</optional>
         </dependency>
 
-        <!-- RabbitMQ for event publishing -->
+        <!-- RocketMQ for event publishing (替代 RabbitMQ) -->
+        <dependency>
+            <groupId>org.apache.rocketmq</groupId>
+            <artifactId>rocketmq-spring-boot-starter</artifactId>
+            
+        </dependency>
+
+        <!-- Redis for caching -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-amqp</artifactId>
+            <artifactId>spring-boot-starter-data-redis</artifactId>
         </dependency>
 
         <!-- Common module (service auth, tracing utils, etc.) -->
@@ -83,6 +87,18 @@
             <groupId>com.example</groupId>
             <artifactId>common</artifactId>
             <version>1.0.0</version>
+        </dependency>
+        
+        <!-- Knife4j OpenAPI3 -->
+        <dependency>
+            <groupId>com.github.xiaoymin</groupId>
+            <artifactId>knife4j-openapi3-jakarta-spring-boot-starter</artifactId>
+        </dependency>
+        
+        <!-- Validation -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-validation</artifactId>
         </dependency>
 
         <!-- Domain module (Result VO, shared DTOs) -->
@@ -94,14 +110,12 @@
 
         <!-- Spring Cloud Sleuth for tracing -->
         <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-starter-sleuth</artifactId>
+            <groupId>io.micrometer</groupId><artifactId>micrometer-tracing-bridge-brave</artifactId>
         </dependency>
 
         <!-- Zipkin reporter -->
         <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-sleuth-zipkin</artifactId>
+            <groupId>io.zipkin.reporter2</groupId><artifactId>zipkin-reporter-brave</artifactId>
         </dependency>
 
         <!-- Sentinel for traffic protection (optional but aligned with other services) -->
@@ -123,6 +137,10 @@
         <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.amqp</groupId>
+            <artifactId>spring-amqp</artifactId>
         </dependency>
     </dependencies>
 
