@@ -21,33 +21,99 @@
         <module>User-server</module>
         <module>Email-server</module>
         <module>Gateway</module>
-        <module>IM-Gateway</module>
+        <module>IM-file-server</module>
+        <module>IM-gateway-netty</module>
         <module>IM-message-server</module>
         <module>IM-relationship-server</module>
-        <module>IM-file-server</module>
         <module>IM-push-server</module>
         <module>IM-av-server</module>
         <module>IM-search-server</module>
         <module>Async-server</module>
+        <module>collab-service</module>
+        <module>admin-service</module>
+        <module>job-service</module>
+        <module>org-service</module>
     </modules>
     <!-- 版本管理 -->
     <properties>
-        <java.version>1.8</java.version>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <!-- Java -->
+        <java.version>21</java.version>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <spring-boot.version>2.7.18</spring-boot.version>
-        <spring-cloud.version>2021.0.8</spring-cloud.version>
-        <spring-cloud-alibaba.version>2021.0.5.0</spring-cloud-alibaba.version>
-        <spring-security-oauth2.version>2.3.4.RELEASE</spring-security-oauth2.version>
-        <spring-security-jwt.version>1.1.1.RELEASE</spring-security-jwt.version>
+        
+        <!-- Spring Boot 3.2.x & Spring Cloud 2023.0.x -->
+        <spring-boot.version>3.2.5</spring-boot.version>
+        <spring-cloud.version>2023.0.3</spring-cloud.version>
+        <spring-cloud-alibaba.version>2023.0.1.2</spring-cloud-alibaba.version>
+        
+        <!-- 数据库 -->
         <mybatis-plus.version>3.5.5</mybatis-plus.version>
+        <mybatis-spring.version>3.0.3</mybatis-spring.version>
         <postgresql.version>42.7.1</postgresql.version>
-        <springfox.version>3.0.0</springfox.version>
-        <knife4j.version>3.0.3</knife4j.version>
+        
+        <!-- API文档 -->
+        <knife4j.version>4.4.0</knife4j.version>
+        
+        <!-- 安全/JWT -->
         <jjwt.version>0.12.5</jjwt.version>
-        <validation-api.version>2.0.1.Final</validation-api.version>
-        <sleuth.version>3.1.9</sleuth.version>
+        
+        <!-- 链路追踪 -->
+        <micrometer-tracing.version>1.2.4</micrometer-tracing.version>
+        
+        <!-- 消息队列 -->
+        <rocketmq.version>2.3.0</rocketmq.version>
+        
+        <!-- 搜索引擎 -->
+        <elasticsearch.version>8.15.0</elasticsearch.version>
+        
+        <!-- 文档解析 -->
+        <tika.version>2.9.2</tika.version>
+        
+        <!-- AI/LLM -->
+        <langchain4j.version>0.33.0</langchain4j.version>
+        
+        <!-- 工具库 -->
+        <hutool.version>5.8.22</hutool.version>
+        <guava.version>33.0.0-jre</guava.version>
+        
+        <!-- Netty -->
+        <netty.version>4.1.109.Final</netty.version>
+        
+        <!-- Protobuf -->
+        <protobuf.version>3.25.1</protobuf.version>
+        
+        <!-- MinIO -->
+        <minio.version>8.5.7</minio.version>
+        
+        <!-- 任务调度 -->
+        <xxl-job.version>2.4.0</xxl-job.version>
+        <quartz.version>2.5.1</quartz.version>
+        
+        <!-- JSON -->
+        <fastjson.version>1.2.83</fastjson.version>
+        
+        <!-- JWT (旧版本兼容) -->
+        <jjwt-old.version>0.11.5</jjwt-old.version>
+        
+        <!-- 序列化 -->
+        <protostuff.version>3.0.0</protostuff.version>
+        <kryo.version>0.6.1</kryo.version>
+        <objenesis.version>3.3</objenesis.version>
+        
+        <!-- Jakarta -->
+        <jakarta.persistence.version>3.1.0</jakarta.persistence.version>
+        <jakarta.validation.version>3.0.2</jakarta.validation.version>
+        
+        <!-- API文档 -->
+        <springdoc.version>2.3.0</springdoc.version>
+        <springfox.version>3.0.0</springfox.version>
+        
+        <!-- Excel -->
+        <easyexcel.version>3.3.3</easyexcel.version>
+        
+        <!-- 拼音 -->
+        <pinyin4j.version>2.5.1</pinyin4j.version>
     </properties>
 
     <!-- 依赖管理 -->
