@@ -1,7 +1,9 @@
 package com.example.auth.service;
 
 import com.example.domain.dto.UserDetailsDTO;
+import com.example.domain.vo.Result;
 import com.example.auth.feign.UserServiceClient;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -20,6 +22,7 @@ import java.util.List;
  * 用户详情服务实现
  * 通过 Feign 调用 User-server 获取用户信息
  */
+@Slf4j
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
     
@@ -35,8 +38,16 @@ public class UserDetailsServiceImpl implements UserDetailsService {
             throw new UsernameNotFoundException("请使用邮箱登录，不支持用户名登录: " + emailOrUsername);
         }
         
-        // 2. 通过邮箱查询用户（只查询 provider='email' 的用户）
-        userDetails = userServiceClient.getUserDetailsByEmail(emailOrUsername).getData();
+        // 2. 通过邮箱查询用户
+        try {
+            Result<UserDetailsDTO> result = userServiceClient.getUserDetailsByEmail(emailOrUsername);
+            if (result != null && result.getData() != null) {
+                userDetails = result.getData();
+            }
+        } catch (Exception e) {
+            log.error("调用用户服务失败: email={}", emailOrUsername, e);
+            throw new UsernameNotFoundException("系统错误，请稍后重试");
+        }
         
         if (userDetails == null) {
             throw new UsernameNotFoundException("用户不存在: " + emailOrUsername);
@@ -48,16 +59,11 @@ public class UserDetailsServiceImpl implements UserDetailsService {
         }
         
         // 4. 检查是否有密码（第三方登录用户没有密码）
-        if (userDetails.getPasswordHash() == null) {
-            throw new UsernameNotFoundException("该账户不支持密码登录，请使用第三方登录: " + emailOrUsername);
+        if (userDetails.getPasswordHash() == null || userDetails.getPasswordHash().isEmpty()) {
+            throw new UsernameNotFoundException("该账户不支持密码登录，请使用第三方登录");
         }
         
-        // 5. 检查登录方式
-        if (!"email".equals(userDetails.getProvider())) {
-            throw new UsernameNotFoundException("该账户使用第三方登录: " + userDetails.getProvider() + " 方式");
-        }
-        
-        // 6. 构建权限
+        // 5. 构建权限
         List<GrantedAuthority> authorities = new ArrayList<>();
         if ("admin".equals(userDetails.getUsername()) || "admin@example.com".equals(userDetails.getEmail())) {
             authorities.add(new SimpleGrantedAuthority("ROLE_ADMIN"));
@@ -65,7 +71,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
             authorities.add(new SimpleGrantedAuthority("ROLE_USER"));
         }
         
-        // 4. 返回 UserDetails（注意：username 使用邮箱）
+        // 6. 返回 UserDetails（注意：username 使用邮箱）
         return org.springframework.security.core.userdetails.User.builder()
                 .username(userDetails.getEmail() != null ? userDetails.getEmail() : userDetails.getUsername())
                 .password(userDetails.getPasswordHash())
