@@ -3,30 +3,29 @@ package com.example.user.service;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.example.domain.dto.UserDetailsDTO;
 import com.example.domain.model.User;
-import com.example.domain.model.UserCredential;
-import com.example.user.feign.RelationshipUserClient;
-import com.example.user.mapper.UserCredentialMapper;
+import com.example.domain.model.UserProfile;
+import com.example.domain.model.PasswordHistory;
 import com.example.user.mapper.UserMapper;
+import com.example.user.mapper.PasswordHistoryMapper;
 import com.example.user.validator.PasswordValidator;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.http.ResponseEntity;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
 /**
- * @author Junjie
- * @version 1.0.0
- * @date 2025-11-06
- * 用户服务
- * 处理用户相关的业务逻辑
+ * 用户核心服务
+ * 基于新数据库设计，处理用户的基本信息管理
+ * 
+ * @author System
+ * @since 2024-11-25
  */
 @Slf4j
 @Service
@@ -36,88 +35,120 @@ public class UserService {
     private UserMapper userMapper;
     
     @Autowired
-    private UserCredentialMapper userCredentialMapper;
+    private PasswordHistoryMapper passwordHistoryMapper;
+    
+    @Autowired
+    private UserProfileService userProfileService;
+    
+    @Autowired
+    private EmailVerificationService emailVerificationService;
     
     @Autowired
     private PasswordEncoder passwordEncoder;
     
     @Autowired
-    private RabbitTemplate rabbitTemplate;
-    
-    @Autowired
     private PasswordValidator passwordValidator;
     
-    @Autowired
-    private RelationshipUserClient relationshipUserClient;
+    private static final int PASSWORD_HISTORY_LIMIT = 5;
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
      * 根据邮箱获取用户详情（用于登录认证）
      */
+    @Cacheable(value = "userDetails", key = "#email")
     public UserDetailsDTO getUserDetailsByEmail(String email) {
         try {
-            log.info("开始查询用户详情: email={}", email);
+            User user = userMapper.selectOne(
+                new QueryWrapper<User>().eq("email", email)
+            );
             
-            // 调试：先查用户
-            User user = userMapper.getUserByEmail(email);
             if (user == null) {
-                log.warn("用户不存在: email={}", email);
+                log.debug("用户不存在: email={}", email);
                 return null;
             }
-            log.info("找到用户: userId={}, email={}", user.getId(), email);
             
-            // 调试：查询凭证
-            UserCredential credential = userCredentialMapper.findByUserIdAndProvider(user.getId(), "email");
-            if (credential == null) {
-                log.warn("用户凭证不存在: userId={}, provider=email", user.getId());
-            } else {
-                log.info("找到用户凭证: credentialId={}, userId={}, provider={}, passwordHash={}", 
-                    credential.getId(), credential.getUserId(), credential.getProvider(),
-                    credential.getPasswordHash() != null ? "存在(长度:" + credential.getPasswordHash().length() + ")" : "NULL");
+            // 构建用户详情DTO
+            UserDetailsDTO dto = new UserDetailsDTO();
+            dto.setUserId(user.getId());
+            dto.setUsername(user.getUsername());
+            dto.setEmail(user.getEmail());
+            dto.setPasswordHash(user.getPasswordHash());
+            dto.setEmailVerified(user.getEmailVerified());
+            
+            // 获取用户详细资料
+            UserProfile profile = userProfileService.getUserProfile(user.getId());
+            if (profile != null) {
+                dto.setDisplayName(profile.getNickname());
+                dto.setAvatarUrl(profile.getAvatarUrl());
             }
             
-            // 正常查询
-            UserDetailsDTO userDetails = userMapper.getUserDetailsByEmail(email);
-            if (userDetails != null) {
-                log.info("查询用户详情成功: userId={}, passwordHash={}", 
-                    userDetails.getUserId(),
-                    userDetails.getPasswordHash() != null ? "存在(长度:" + userDetails.getPasswordHash().length() + ")" : "NULL");
-            } else {
-                log.warn("getUserDetailsByEmail 返回 null: email={}", email);
-            }
-            return userDetails;
+            return dto;
         } catch (Exception e) {
-            log.error("查询用户详情失败: email={}", email, e);
+            log.error("获取用户详情失败: email={}", email, e);
             return null;
         }
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
+     * 根据用户ID获取用户详情
+     */
+    @Cacheable(value = "userDetails", key = "'id:' + #userId")
+    public UserDetailsDTO getUserDetailsById(Long userId) {
+        try {
+            User user = userMapper.selectById(userId);
+            if (user == null) {
+                return null;
+            }
+            
+            UserDetailsDTO dto = new UserDetailsDTO();
+            dto.setUserId(user.getId());
+            dto.setUsername(user.getUsername());
+            dto.setEmail(user.getEmail());
+            dto.setPasswordHash(user.getPasswordHash());
+            dto.setEmailVerified(user.getEmailVerified());
+            
+            // 获取用户详细资料
+            UserProfile profile = userProfileService.getUserProfile(userId);
+            if (profile != null) {
+                dto.setDisplayName(profile.getNickname());
+                dto.setAvatarUrl(profile.getAvatarUrl());
+            }
+            
+            return dto;
+        } catch (Exception e) {
+            log.error("获取用户详情失败: userId={}", userId, e);
+            return null;
+        }
+    }
+    
+    /**
      * 根据用户名获取用户详情
      */
+    @Cacheable(value = "userDetails", key = "'username:' + #username")
     public UserDetailsDTO getUserDetailsByUsername(String username) {
         try {
-            return userMapper.getUserDetailsByUsername(username);
+            User user = userMapper.selectOne(
+                new QueryWrapper<User>().eq("username", username)
+            );
+            
+            if (user == null) {
+                return null;
+            }
+            
+            return getUserDetailsById(user.getId());
         } catch (Exception e) {
-            log.error("查询用户详情失败: username={}", username, e);
+            log.error("获取用户详情失败: username={}", username, e);
             return null;
         }
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
      * 检查邮箱是否存在
      */
     public boolean checkEmailExists(String email) {
         try {
-            int count = userMapper.checkEmailExists(email);
+            Long count = userMapper.selectCount(
+                new QueryWrapper<User>().eq("email", email)
+            );
             return count > 0;
         } catch (Exception e) {
             log.error("检查邮箱失败: email={}", email, e);
@@ -126,14 +157,13 @@ public class UserService {
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
      * 检查用户名是否存在
      */
     public boolean checkUsernameExists(String username) {
         try {
-            int count = userMapper.checkUsernameExists(username);
+            Long count = userMapper.selectCount(
+                new QueryWrapper<User>().eq("username", username)
+            );
             return count > 0;
         } catch (Exception e) {
             log.error("检查用户名失败: username={}", username, e);
@@ -144,19 +174,24 @@ public class UserService {
     /**
      * 更新用户最后登录时间
      */
+    @CacheEvict(value = "userDetails", allEntries = true)
     public void updateLastLoginTime(String email) {
         try {
-            userMapper.updateLastLoginTime(email);
-            log.info("更新最后登录时间成功: email={}", email);
+            User user = userMapper.selectOne(
+                new QueryWrapper<User>().eq("email", email)
+            );
+            
+            if (user != null) {
+                user.setLastLoginAt(LocalDateTime.now());
+                userMapper.updateById(user);
+                log.info("更新最后登录时间成功: email={}", email);
+            }
         } catch (Exception e) {
             log.error("更新最后登录时间失败: email={}", email, e);
         }
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
      * 根据用户ID获取用户信息
      */
     public User getUserById(Long userId) {
@@ -169,9 +204,20 @@ public class UserService {
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
+     * 根据邮箱获取用户信息
+     */
+    public User getUserByEmail(String email) {
+        try {
+            return userMapper.selectOne(
+                new QueryWrapper<User>().eq("email", email)
+            );
+        } catch (Exception e) {
+            log.error("查询用户失败: email={}", email, e);
+            return null;
+        }
+    }
+    
+    /**
      * 创建新用户（邮箱注册）
      */
     @Transactional(rollbackFor = Exception.class)
@@ -194,37 +240,28 @@ public class UserService {
             User user = new User();
             user.setUsername(username);
             user.setEmail(email);
+            user.setPasswordHash(passwordEncoder.encode(password)); // 新数据库设计：密码直接存在users表
             user.setEmailVerified(false);
+            user.setStatus("active");
             user.setCreatedAt(LocalDateTime.now());
             user.setUpdatedAt(LocalDateTime.now());
             
             userMapper.insert(user);
             log.info("用户创建成功: userId={}, username={}, email={}", user.getId(), username, email);
             
-            // 3. 创建用户凭证（密码）
-            log.info("准备创建用户凭证: userId={}, provider=email", user.getId());
-            UserCredential credential = new UserCredential();
-            credential.setUserId(user.getId());
-            credential.setProvider("email");
-            String encodedPassword = passwordEncoder.encode(password);
-            credential.setPasswordHash(encodedPassword);
-            credential.setCreatedAt(LocalDateTime.now());
+            // 4. 创建默认用户详情
+            userProfileService.getUserProfile(user.getId()); // 会自动创建默认profile
             
-            log.info("密码已加密: userId={}, passwordHashLength={}", user.getId(), encodedPassword != null ? encodedPassword.length() : 0);
+            // 5. 记录密码历史
+            savePasswordHistory(user.getId(), user.getPasswordHash());
             
-            int insertRows = userCredentialMapper.insert(credential);
-            log.info("用户凭证创建成功: userId={}, insertRows={}, credentialId={}", 
-                user.getId(), insertRows, credential.getId());
-            
-            // 4. 发送欢迎邮件（异步）
-            sendWelcomeEmail(email, username);
-
-            // 5. 在 Neo4j 中初始化用户节点（关系服务）
+            // 6. 发送验证邮件
             try {
-                relationshipUserClient.ensureUserNode(user.getId());
-                log.info("已通知关系服务在 Neo4j 中初始化用户节点: userId={}", user.getId());
+                emailVerificationService.sendVerificationCode(email);
+                log.info("验证邮件已发送: email={}", email);
             } catch (Exception e) {
-                log.warn("通知关系服务初始化 Neo4j 用户节点失败: userId={}", user.getId(), e);
+                log.warn("发送验证邮件失败，但用户已创建: email={}", email, e);
+                // 不影响注册流程
             }
             
             return user;
@@ -236,70 +273,43 @@ public class UserService {
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
-     * 创建或更新第三方登录用户
+     * 记录密码历史
      */
-    @Transactional(rollbackFor = Exception.class)
-    public User createOrUpdateOAuthUser(String provider, String providerUserId, 
-                                       String username, String email, String avatarUrl) {
+    @Async
+    private void savePasswordHistory(Long userId, String passwordHash) {
         try {
-            // 1. 查找是否已存在
-            User existingUser = userMapper.findByProviderAndProviderUserId(provider, providerUserId);
+            PasswordHistory history = new PasswordHistory();
+            history.setUserId(userId);
+            history.setPasswordHash(passwordHash);
+            history.setCreatedAt(LocalDateTime.now());
+            passwordHistoryMapper.insert(history);
             
-            if (existingUser != null) {
-                // 更新用户信息
-                existingUser.setUsername(username);
-                existingUser.setEmail(email);
-                existingUser.setUpdatedAt(LocalDateTime.now());
-                userMapper.updateById(existingUser);
-
-                try {
-                    relationshipUserClient.ensureUserNode(existingUser.getId());
-                    log.info("已通知关系服务在 Neo4j 中初始化用户节点(第三方用户-更新): userId={}", existingUser.getId());
-                } catch (Exception e) {
-                    log.warn("通知关系服务初始化 Neo4j 用户节点失败(第三方用户-更新): userId={}", existingUser.getId(), e);
+            // 删除超过限制的旧记录
+            List<PasswordHistory> histories = passwordHistoryMapper.selectList(
+                new QueryWrapper<PasswordHistory>()
+                    .eq("user_id", userId)
+                    .orderByDesc("created_at")
+            );
+            
+            if (histories.size() > PASSWORD_HISTORY_LIMIT) {
+                for (int i = PASSWORD_HISTORY_LIMIT; i < histories.size(); i++) {
+                    passwordHistoryMapper.deleteById(histories.get(i).getId());
                 }
-
-                log.info("更新第三方用户: userId={}, provider={}", existingUser.getId(), provider);
-                return existingUser;
             }
-            
-            // 2. 创建新用户
-            User user = new User();
-            user.setUsername(username);
-            user.setEmail(email);
-            user.setEmailVerified(true); // 第三方登录默认邮箱已验证
-            user.setCreatedAt(LocalDateTime.now());
-            user.setUpdatedAt(LocalDateTime.now());
-            
-            userMapper.insert(user);
-            log.info("创建第三方用户: userId={}, provider={}", user.getId(), provider);
-            
-            // 3. 创建用户凭证
-            UserCredential credential = new UserCredential();
-            credential.setUserId(user.getId());
-            credential.setProvider(provider);
-            credential.setProviderUserId(providerUserId);
-            credential.setCreatedAt(LocalDateTime.now());
-            
-            userCredentialMapper.insert(credential);
-            log.info("创建第三方凭证: userId={}, provider={}", user.getId(), provider);
-
-            try {
-                relationshipUserClient.ensureUserNode(user.getId());
-                log.info("已通知关系服务在 Neo4j 中初始化用户节点(第三方用户-创建): userId={}", user.getId());
-            } catch (Exception e) {
-                log.warn("通知关系服务初始化 Neo4j 用户节点失败(第三方用户-创建): userId={}", user.getId(), e);
-            }
-
-            return user;
-            
         } catch (Exception e) {
-            log.error("创建或更新第三方用户失败: provider={}, providerUserId={}", provider, providerUserId, e);
-            throw new RuntimeException("创建或更新第三方用户失败", e);
+            log.error("保存密码历史失败: userId={}", userId, e);
         }
+    }
+    
+    /**
+     * 创建或更新OAuth用户（由OAuthService处理）
+     */
+    @Deprecated
+    public User createOrUpdateOAuthUser(String provider, String providerUserId,
+                                       String username, String email, String avatarUrl) {
+        // 此功能已迁移到 OAuthService
+        log.warn("调用已废弃的方法createOrUpdateOAuthUser，请使用OAuthService");
+        return null;
     }
     
     /**
@@ -329,37 +339,51 @@ public class UserService {
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
      * 修改密码
      */
     @Transactional(rollbackFor = Exception.class)
+    @CacheEvict(value = "userDetails", allEntries = true)
     public boolean changePassword(Long userId, String oldPassword, String newPassword) {
         try {
-            // 1. 验证旧密码
-            UserCredential credential = userCredentialMapper.findByUserIdAndProvider(userId, "email");
-            if (credential == null) {
-                log.warn("用户凭证不存在: userId={}", userId);
+            User user = userMapper.selectById(userId);
+            if (user == null) {
+                log.warn("用户不存在: userId={}", userId);
                 return false;
             }
             
-            if (!passwordEncoder.matches(oldPassword, credential.getPasswordHash())) {
+            // 1. 验证旧密码
+            if (!passwordEncoder.matches(oldPassword, user.getPasswordHash())) {
                 log.warn("旧密码错误: userId={}", userId);
                 return false;
             }
             
-            // 2. 更新密码
+            // 2. 验证新密码强度
+            String passwordError = passwordValidator.validate(newPassword);
+            if (passwordError != null) {
+                log.warn("新密码强度不足: userId={}, error={}", userId, passwordError);
+                throw new IllegalArgumentException(passwordError);
+            }
+            
+            // 3. 检查密码历史
+            if (checkPasswordHistory(userId, newPassword)) {
+                log.warn("密码与最近使用的密码重复: userId={}", userId);
+                throw new IllegalArgumentException("不能使用最近使用过的密码");
+            }
+            
+            // 4. 更新密码
             String newPasswordHash = passwordEncoder.encode(newPassword);
-            int rows = userCredentialMapper.updatePassword(userId, newPasswordHash);
+            user.setPasswordHash(newPasswordHash);
+            user.setUpdatedAt(LocalDateTime.now());
+            int rows = userMapper.updateById(user);
             
             if (rows > 0) {
+                // 记录密码历史
+                savePasswordHistory(userId, newPasswordHash);
                 log.info("修改密码成功: userId={}", userId);
                 return true;
             }
             
             return false;
-            
         } catch (Exception e) {
             log.error("修改密码失败: userId={}", userId, e);
             return false;
@@ -367,31 +391,27 @@ public class UserService {
     }
     
     /**
-     * @author Junjie
-     * @version 1.0.0
-     * @date 2025-11-06
-     * 发送欢迎邮件
+     * 检查密码历史
      */
-    private void sendWelcomeEmail(String email, String username) {
+    private boolean checkPasswordHistory(Long userId, String newPassword) {
         try {
-            Map<String, String> emailData = new HashMap<>();
-            emailData.put("to", email);
-            emailData.put("subject", "欢迎注册");
-            emailData.put("content", String.format("欢迎 %s 注册我们的平台！", username));
+            List<PasswordHistory> histories = passwordHistoryMapper.selectList(
+                new QueryWrapper<PasswordHistory>()
+                    .eq("user_id", userId)
+                    .orderByDesc("created_at")
+                    .last("LIMIT " + PASSWORD_HISTORY_LIMIT)
+            );
             
-            // 使用交换机和路由键发送消息
-            rabbitTemplate.convertAndSend("user.exchange", "email.welcome", emailData);
-            log.info("欢迎邮件已发送到交换机: email={}, exchange=user.exchange, routingKey=email.welcome", email);
-            
+            for (PasswordHistory history : histories) {
+                if (passwordEncoder.matches(newPassword, history.getPasswordHash())) {
+                    return true;
+                }
+            }
+            return false;
         } catch (Exception e) {
-            log.error("发送欢迎邮件失败: email={}", email, e);
+            log.error("检查密码历史失败: userId={}", userId, e);
+            return false;
         }
     }
     
-    /**
-     * 根据邮箱获取用户
-     */
-    public User getUserByEmail(String email) {
-        return userMapper.getUserByEmail(email);
-    }
 }
