@@ -3,11 +3,11 @@ import { createPinia } from 'pinia'
 import ElementPlus from 'element-plus'
 import 'element-plus/dist/index.css'
 import * as ElementPlusIconsVue from '@element-plus/icons-vue'
-import Antd from 'ant-design-vue'
-import 'ant-design-vue/dist/reset.css'
-import * as AntdIcons from '@ant-design/icons-vue'
+import zhCn from 'element-plus/dist/locale/zh-cn.mjs'
+
 import App from './App.vue'
 import router from './router'
+import './styles/global.css'
 
 const app = createApp(App)
 
@@ -16,13 +16,8 @@ for (const [key, component] of Object.entries(ElementPlusIconsVue)) {
   app.component(key, component)
 }
 
-for (const [key, component] of Object.entries(AntdIcons)) {
-  app.component(key, component)
-}
-
 app.use(createPinia())
 app.use(router)
-app.use(ElementPlus)
-app.use(Antd)
+app.use(ElementPlus, { locale: zhCn })
 
 app.mount('#app')
